+++ MMSEngine/src/MMSEngineDBFacade.h
@@ -1797,7 +1797,7 @@ class MMSEngineDBFacade
 
 	void updateEncodingJobProgress(int64_t encodingJobKey, double encodingPercentage);
 
-	void updateEncodingRealTimeInfo(int64_t encodingJobKey, int encodingPid, long realTimeFrameRate, long realTimeBitRate);
+	void updateEncodingRealTimeInfo(int64_t encodingJobKey, int encodingPid, long realTimeFrameRate, double realTimeBitRate);
 
 	bool updateEncodingJobFailuresNumber(int64_t encodingJobKey, long failuresNumber);
